+++ src/main/java/uz/bookshop/domain/dto/response_dto/OrderDetailsResponseDTO.java
@@ -5,8 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigInteger;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -16,5 +14,5 @@ public class OrderDetailsResponseDTO {
     private Long id;
     private BookResponseDTO book;
     private Integer quantity;
-    private BigInteger totalPrice;
+    private Integer totalPrice;
 }
